Add testcontainers to database tests
